+++ common/misc/config.h
@@ -24,7 +24,7 @@ public:
    class TileParameters
    {
    public:
-      TileParameters(std::string core_type, volatile float frequency, volatile float voltage, std::string l1_icache_type, std::string l1_dcache_type, std::string l2_cache_type):
+      TileParameters(std::string core_type, volatile double frequency, volatile double voltage, std::string l1_icache_type, std::string l1_dcache_type, std::string l2_cache_type):
          m_core_type(core_type),
          m_frequency(frequency),
          m_voltage(voltage),
@@ -35,16 +35,16 @@ public:
       ~TileParameters() {}
 
       std::string getCoreType() { return m_core_type; }
-      volatile float getFrequency() { return m_frequency; }
-      volatile float getVoltage() { return m_voltage; }
+      volatile double getFrequency() { return m_frequency; }
+      volatile double getVoltage() { return m_voltage; }
       std::string getL1ICacheType() { return m_l1_icache_type; }
       std::string getL1DCacheType() { return m_l1_dcache_type; }
       std::string getL2CacheType() { return m_l2_cache_type; }
    
    private:
       std::string m_core_type;
-      volatile float m_frequency;
-      volatile float m_voltage;
+      volatile double m_frequency;
+      volatile double m_voltage;
       std::string m_l1_icache_type;
       std::string m_l1_dcache_type;
       std::string m_l2_cache_type;
@@ -53,17 +53,17 @@ public:
    class NetworkParameters
    {
    public:
-      NetworkParameters(std::string type, volatile float frequency):
+      NetworkParameters(std::string type, volatile double frequency):
          m_type(type), m_frequency(frequency)
       {}
       ~NetworkParameters() {}
 
       std::string getType() { return m_type; }
-      volatile float getFrequency() { return m_frequency; }
+      volatile double getFrequency() { return m_frequency; }
    
    private:
       std::string m_type;
-      volatile float m_frequency;
+      volatile double m_frequency;
    };
 
    enum SimulationMode
@@ -154,8 +154,8 @@ public:
 
    // Tile & Network Parameters
    std::string getCoreType(tile_id_t tile_id);
-   volatile float getTileFrequency(tile_id_t tile_id);
-   volatile float getTileVoltage(tile_id_t tile_id);
+   volatile double getTileFrequency(tile_id_t tile_id);
+   volatile double getTileVoltage(tile_id_t tile_id);
    std::string getL1ICacheType(tile_id_t tile_id);
    std::string getL1DCacheType(tile_id_t tile_id);
    std::string getL2CacheType(tile_id_t tile_id);
