+++ common/tile/memory_subsystem/pr_l1_pr_l2_dram_directory_msi/dram_directory_cntlr.cc
@@ -17,8 +17,8 @@ DramDirectoryCntlr::DramDirectoryCntlr(MemoryManager* memory_manager,
       string dram_directory_type_str,
       string dram_directory_access_time_str,
       UInt32 num_dram_cntlrs,
-      float frequency,
-      float voltage)
+      double frequency,
+      double voltage)
    : _memory_manager(memory_manager)
    , _dram_cntlr(dram_cntlr)
 {
