+++ common/tile/memory_subsystem/pr_l1_sh_l2_msi/l2_cache_cntlr.h
@@ -36,8 +36,8 @@ namespace PrL1ShL2MSI
                    string L2_cache_replacement_policy,
                    UInt32 L2_cache_access_delay,
                    bool L2_cache_track_miss_types,
-                   float frequency,
-                   float voltage);
+                   double frequency,
+                   double voltage);
       ~L2CacheCntlr();
 
       Cache* getL2Cache() { return _L2_cache; }
